Started work on the text module.
A Talkbox is a dialogue box with a Portrait and a Text, rendered in a
particular Font.
A Portrait will be an image displayed to show who is talking.
A Text is a char array which will be displayed as text on screen.
A Font is an image and related metadata describing how to convert a char
into a glyph for display.

(Minor: Renamed text.png to fixedsys.png to reflect its font of origin.)
(Minor: Added fontinfo.txt to save font information.)