Initial Jumping Commit
Added Actions:
-Jump

The hero now jumps on key input x when facing a space with a height
difference of 1.  There is currently no animation (so it does a funky
walking thing), and the transition is only implemented for left and
right jumps (up and down jumps are currently instantaneous).

Also updated the rendering order to fix a previously unforeseen
occlusion issue.