Player Camera Redesign
Redesigned the camera so that it now works smoothly with jumping at
display edges.

Removed Action: 
	StepCamera, as it is no longer necessary.

Still to implement for camera:
	1) Snapping to correct position on teleportation to edge (mostly for
map entry, but this should be good for robustness regardless)
	2) update the isBlocking method to set the permanently off-camera
parts of the top and bottom map edges to be blocking even if they are
free.

Further Notes:
	the house has some strange display glitches on the back wall and
any roof adjacent to a wall, look into this stuff...