Player Camera Redesign Part II
Added previously missing functionality for dealing with camera locking
at the edges.

New functionality:
	1) Edges snap to correct position during high-speed movement across
lock borders; adjusted bottom border to stop at the edge of the top
height slice; camera snaps to correct position when teleporting into a
lock zone.
	2) Small maps are supported again by snapping the camera to the center
of the cube if it is too small for camera scrolling.
	3) Updated isBlocked() method in the world class to block off-screen
edges at various heights caused by the top/bottom screen locking.

Removed:
	Clockwise and counter clockwise rotation for maps, this was old
functionality that ceased to function correctly with the addition of
agents and I forgot to remove it before now!