Heralding the Arrival of Talldorf
With the thing graphical tests, it became apparent that Baldorf was a
bit too small for his world.  As such, he's now two pixels taller, and
fits in much better.

Also his coat was swinging the wrong way on some of his steps, and so
the animation has been fixed to better follow the laws of physics.