Tables and Chairs
Added some new things, as well as interaction functionality for each
thing/agent (defaulted to nothing), activated by pressing z when facing
a thing/agent to interact with.  An example of which is that chairs and
stools can be pulled out or pushed in through interaction.  When pushed
in, the chair can't be stepped on, but when pulled out an agent can step
on a chair to "sit" in it.

New Things:
	Chair - standard wooden chair
	Stool - standard wooden stool
	Table - round table, occupies one grid cell

New Textures:
	Thing2 - chairs, tables, and candles