Candles and Light Source Improvements
Refactored light sources to work more consistently and efficiently,
things now have an update that works similarly to the updating of agents
(shown by the Candle class which animates the flame and causes light
flickering), reworked blocking and crossability for terrain and objects
to work as was originally intended, and added some new candle objects to
give light sources some much needed graphics.

New Things:
* Candle - a standard candle meant to sit at table height
* WallCandle - a wall-mounted candle (some work is still needed for
correct interaction between the walls and things)

Changes to allow multiple things to exist within each cell are coming
next - allowing the candle to actually sit on a table as intended.