Thing Stacks are Go!
Updated the world's representation of things to use the new
ThingGridCell class, which allows more than one thing to be exist within
the same grid cell.  Actions are also updated to reflect this change. 
The demo map is also updated to show some stacks of things.  Note that
ramp interaction with things stacked on top may not be correct - this is
a special case anyway that may not be allowed in the first place, but
will be updated if necessary.

Also, forgot to mention previously in the Candles and Light Source
Improvements commit, I added the functionality for "anti lights" - i.e.,
objects that absorb light instead of emit light - just for the fun of
it.

New Things:
* ThingGridCell - class to contain multiple things within one cell