All Things Fireplace
Added lightable fireplaces which will burn out over time (can be
refilled with wood and relit by interacting), fireplace sides and
chimneys for aesthetic purposes, and firewood, which is the first
resource-type thing, meaning it will be diminished as it is taken.
 This can currently be tested by interacting with the firewood, which
is set to contain 15 pieces of wood, and will update its texture at
intervals of 5.

Also added vertical chairs and made minor adjustments to the positioning
of some objects.

New Things:
* Fireplace - a light source whose brightness depends on the amount of
wood it contains, which will burn down overtime
* FireplaceSide - the sides for the core Fireplace, blocking and
inanimate
* FireplaceChimney - the top part of a Fireplace, can be open for the
top or continuously loop upon itself to extend the height, otherwise
blocking and inanimate
* Firewood - wood resource that depletes as it is taken