New Graphics and Text Framework
Added graphics for a new character, Innesley, to go with Baldorf.

Also added a lot of text stuff including graphics for a font and a
textbox, and a set of functions under utils for breaking strings of text
up into individual textboxes with appropriate line lengths based on how
the font will fit into the text box.