Textboxes!
Implemented textbox display for arbitrary phrases.  The text boxes will
display text slowly, with pauses on certain punctuation, at a
dialogue-mimicing speed.  Holding the 'c' key will significantly speed
up text display, for anyone in a hurry (or anyone who wants to quickly
replay parts of a conversation).  When a text box is full, it will not
advance to the next text box until either the 'z' or 'x' key is pressed.

The test world currently starts with a text box display text.  The
actual action for interacting with a person/thing displaying a text box
is still being implemented!