Textboxes II
Added name display over textboxes.  This will always be visible,
initialize to "..." if a name is unspecified, and size dynamically to
the name length.  There's currently no upper limit.

Looking closely at the text itself, some of the textures are off by a
pixel.  This should be addressed in the next bug fix.