Agent Interaction
Tuned and tested agent interaction.  This can be seen by using the z key
to interact with the character in the inn.  Agents can also be
interacted with by other non-player agents to produce different
behavior, see the new Innkeeper.java class for an example.

New Agents:
* Innkeeper - currently just talks with the player character

New Actions:
* Say - present text in textboxes being said by a single character; also
can optionally turn the agent towards the hero before speaking

The Say action will likely need adjustment to work with moving agents,
but for now it works well with stationary agents.