Consistent Step+Interact
Finally came up with a solution for handling upwards stepping so that
once the step is initiated, the grid location of the character is one
step forward, with no incorrect overlapped rendering.  The solution
involves some shenannigans with having the Placeholder agent render the
original agent, so Placeholders are now tied directly to the agent they
are standing in for.

This allows consistent stepping and interacting from the perspective of
the player character, and it provides a cleaner solution to some of the
annoying delayed shadowing behavior of upward stepping.  These changes
will of course also affect non-player characters and their interacting.

Note that this functionality has not been added to the Jump or RampStep
ctions.  The upwards RampStep is inconsistent with the sideways ramp
step anyway, but something like this should probably be changed for the
Jump action, to prevent interacting with a character that has just left
the ground.