Action Interruption
Agents now maintain a queue of actions which can be pushed to if an
outside entity interrupts their current action.  As a result, all
actions must now have two more methods implemented to denote whether the
current action can be interrupted, and if so, whether they are currently
at an interruptible point.

This can be seen by talking to the (now pacing) Innkeeper, where an
interaction request will not launch a conversation until the Innkeeper's
step action is finished.  When the conversation is finished, she will
continue on with the pacing wherever it left off.