Thing Refactoring and Table Talk
Refactored a lot of things to use a more consistent representation and
to cut down on the number of repetitive classes.  Many of the more
complex things are also now instantiated using a Builder design pattern,
so optional parameters are a lot more sensible.

Things Refactored:
* HorizontalBar, VerticalBar -> Bar
* Stairs, StairsIndoor -> Stairs
* Chair, Stool -> Chair
* StairsIndoorBottom -> StairsBottom

Things also now have a remove() method, which will allow them to make
any necessary changes to other stuff around them, e.g. removing an piece
of Stairs will also remove the StairsIndoorBottom piece associated with
it.

As for new functionality, the Table and Bar can now be talked across, as
seen in the demo inn.  Just use the interaction (z) button.