A* Path Planning I
I implemented an A* path planner that will plan a path using only the
SimpleStep actions. There needs to be some sort of cap on how long it
will run, however, because if a path can't be found on a large map it
will take a while to realize it.  The planner can be used with the new
WalToPoint action.

Some stuff was fixed up with FollowPath as well, so non-looping paths
should work correctly.

New Actions:
* WalkToPoint - plan and execute a path to a given point while avoiding
all obstacles