Wooden Beams and Shadowed Fireplaces
All of the various Beams have been refactored into a single class called
Beam, using the factory pattern similrly to the Bar thing.  Diagonal
beams still need a connectionContext, but otherwise they seem to be
working well.  Things also have a few new attributes that allow them to
be rendered as black boxes when they are at the height limit, see the
fireplace for an example.

Things Refactored:
* *Beam*.java -> Beam