Consistency Fixes
Added connection context to diagonal beams, which fixed a rendering
problem on them, changed wall candles to exist on the appropriate square
so they're more consistent with beams, adjusted lighting on in-wall
(i.e. on-wall) things so that they are not shadowed (as opposed to
always shadowed) so they are now consistent with the lighting on the
wall, and fixed a shadowing issue on upwards steps.