Massive Action Refactor and Incremental Path Planning
Changed the way actions work so that their state can be maintained much
easier.  Actions no longer take a parameter of Strings on the execute
method.  Instead, state is set when the action is constructed.  The
general flow now is that an action gets constructed, it is then executed
until it is complete, and it won't do anything else.  To perform the
same action again, or to perform a new action, a new instance of the
aciton can be constructed.  This makes interrupting actions much easier,
and overall the state is much more clearly maintained within the action.

Actions Refactored:
 * All of them, significantly!

Path planning has also been changed so that the planner itself is now an
object.  It is set up to iteratively plan in small chunks, so that no
single agent's planner will slow down the execution loop with a long
planning time.  The paramters are configurable within the planning
class.  Globally useful planning functions remain and were refactored
into the PathPlannerUtils class.

New Planner:
 * AStar - basic AStar planner where movement type can be set to change
the action domain (currently only implemented for SimpleStepping)
 
Outstanding Bugs:
 * Jump - jumping up while facing up seems to leave behind a placeholder
ghost tied to the jumping agent, which will reappear on all subsequent
forward-facing-up-jumps...