Vertical Ramps
Reworked vertical ramps so they appear to work like horizontal ramps.
In reality it's taking advantage of impossible physical space, but with
the in-game camera angle I think the result looks pretty good.  The
ramps themselves work analogously to ladders, where the character is
moving up in the z direction rather than up in the z and forward in the
y.  I also turned off Innesley for now, as the vertical ramps aren't in
the AStar planner yet, and I didn't want to confuse it too much!

New Actions
 * VerticalRampStep - perform steps up, down, or even across vertical
ramps.
 
Refactored Actions
 * RampStep -> HorizontalRampStep - the functionality was different
enough that I broke the rampStep action into two parts based on vertical
or horizontal ramp use.  This also added extra functions to the
PathPlnnerUtils class, also denoted with the vertical or horizontal
name.

Refactored Things
 * Stairs - The indoor wooden stairs can now be placed vertically.  The
outdoor wooden stairs won't work correctly if you try to use them
vertically, however, and will likely be removed in the future.