More Support for Multiple Movement Classes
Updated FollowPath to allow specification of a movement class, so an
agent using it can be limited to only simple stepping (no stairs, etc.),
or unconstrained using stepping (regular steps and stairs).

Agents also can be rendered one layer higher when a height max is
enforced due to occlusion, which makes everything look a lot nicer.