Climbing I
Implemented climbing actions for agents to traverse climbing surfaces.
Also, fixed a path planning bug for stairs that could result in
impossible to execute paths on the bottom of horizontal stairs.  The
climbing action still needs its own animation, and the path planner does
not yet include it.

New Actions:
 * Climb - a single-step climb up, down, or accross a climbing surface,
such as a ladder or rope.
 * StepOrClimb - a new higher-level movement option that will either
execute a simple step, a ramp step, or a climb depending on what is most
appropriate for the environment being moved onto.