Cross Section Textures
Initial support of cross section textures to replace the black box that
was rendered when the screen cut off at the player's height.  For now
they're based on a four-connected check, so inside corners don't look
quite right.  Also note that the inn looks a bit funky, but that's more
because it wasn't designed with this in mind!