Cross Sections II
Switched the terrain type used for cross section rendering to that of
the terrain one space below what is being cut off (i.e. the piece of
terrain that you would see the "top" of if everything above it was cut
off), rather than the terrain type of the cut off cell.  The exception
is if the one-below cell is transparent (e.g. a window), in which case
the terrain type of the cut off cell is used instead.

Also updated textures for the filling in the corner gaps left in certain
cross section configurations.  Actually rendering the corners correctly
is still in progress!