Thing Saving and Loading
Thing saving and loading is finished, and works at the map level.  This
also resulted in some cleanup of associated things.  Now a thing (or an
agent) can have a list of associated things (or agents), meaning that if
the master thing gets removed, so do it's associated things.  For
example, see Stairs and StairsBottom, which provides an associated stair
bottom object for visual blending.