Agent Saving and Loading
Agent saving and loading is finished, and works at the map level.  It
doesn't go into so much detail as saving the full state of the agent,
but it will save the position/direction and any state needed for
decision making (so what this mainly leaves out is the currently
executing action, but the agent's behavior should be suitable to decide
what action to execute next reasonably enough anyway).

The position on sideways (left-right) stairs has also been adjusted so
the character doesn't overlap with the edge as much.  This also
necessitated an adjustment to the overall world render order, which has
introduced a new graphical glitch in vertical (up-down in the y-axis)
jumping, but this should resolve itself in the future when the
placeholders are correctly set for jumping anyway.

The only thing that still needs to have saving and loading implemented
is game state information, which will include player position, time of
day, etc.